Rename ff_components
